Removed `Void` variant from `middle::Type` enum.
`Void` is not valid as a type for arguments or in a struct, so it was
removed from the `Type` enum. `middle::Cif` now accepts an
`Option<Type>` as the return argument, and a `None` here signifies a
`Void`.

In addition, a few changes were done to doctests after some more
thorough linting.