Changed `low::Callback*` type signatures
Callback types previously accepted (mutable) references, however the
values originate from pointers managed outside of Rust. By changing the
references to pointers, it is made more clear that special care need to
be taken when accessing the values. In particular, there is no guarantee
for the Mut variants that there are not two mutable references to the
same data at the same time.

The `cif` parameter is kept as a reference as it is read-only and should
not change.