Fix `generate_miri_closure_test_for_ty!` macro
In the last case of `generate_miri_closure_test_for_ty!`, with multiple
arguments and a return value, the closure was invoked twice. Once with
the result pointing to a NULL pointer and once with the actual result
pointer. This was an error, and the closure should only be invoked once
with a pointer to the result area.

In addition, pointers to a null pointer in the macro were exchanged for
NULL pointers to catch anyone trying to read or write to the pointers,
which was the intention from the start.