Add test for default ABI
Libffi v3.5.0 added a new `ffi_get_default_abi` function that returns
the default ABI. A new test was added to ensure that the default abi set
by libffi-rs is identical to the one used when compiling libffi.

This test is disabled when using the system libffi as it is not
currently required to use libffi v3.5.0.