Fix default ABI on powerpc64
The `LINUX_STRUCT_ALIGN` was not set on big endian powerpc64, causing
the default ABI to be set incorrectly when compiling for powerpc64.

This was discovered due to the new test for default ABIs that utilize
`ffi_get_default_abi`, which returns the default ABI that is used when
compiling libffi.