Fix `ffi_closure` (#43)
* Fix align of `ffi_trampoline`

`ffi_trampoline` had an align of 64 for some unknown reason. This commit
removes the align to help make sure that the layout of `ffi_trampoline`
matches the layout used by libffi.

* Fix `ffi_closure` layout

On Apple systems using ARM CPUs, there is no trampoline in libffi's
`ffi_closure`. Instead there are two pointers `trampoline_table` and
`trampoline_table_entry`. This commit uses conditional compilation to
make sure that the layout of `ffi_closure` is correct on all platforms.

The test to make sure that `ffi_closure` is large enough was moved
from `libffi::low` to `libffi-sys`. The new test now verifies that the size of
the `ffi_closure`s are equal.