Added GitHub action to upgrade libffi
Added a GitHub action to upgrade libffi from libffi's release .tar.gz
file. It creates a new branch, unpacks libffi and generates headers used
when compiling with MSVC. It creates a draft pull request with the
changes so that PR checks are executed when the PR is marked as ready
for review.